+++ s/demo/views/quartz-tripler.ts
@@ -2,7 +2,7 @@
 import {slate} from "../frontend.js"
 import {html} from "../../nexus/html.js"
 
-export const QuartzTripler = slate.light_view(use => (start: number) => {
+export const QuartzTripler = slate.lightView(use => (start: number) => {
 	use.name("quartz-tripler")
 
 	// react hooks state